refactoring solvers to make room for NewtonSolver
