+++ include/ilqgames/solver/ilq_flat_solver.h
@@ -48,6 +48,7 @@
 #include <ilqgames/cost/player_cost.h>
 #include <ilqgames/dynamics/multi_player_flat_system.h>
 #include <ilqgames/solver/game_solver.h>
+#include <ilqgames/solver/ilq_solver.h>
 #include <ilqgames/solver/lq_feedback_solver.h>
 #include <ilqgames/solver/lq_solver.h>
 #include <ilqgames/solver/solver_params.h>
@@ -64,12 +65,12 @@
 
 namespace ilqgames {
 
-class ILQFlatSolver : public GameSolver {
+class ILQFlatSolver : public ILQSolver {
  public:
   virtual ~ILQFlatSolver() {}
   ILQFlatSolver(const std::shared_ptr<Problem>& problem,
                 const SolverParams& params = SolverParams())
-    : GameSolver(problem, params) {
+      : ILQSolver(problem, params) {
     // Precompute linearization.
     CHECK(problem_->Dynamics()->TreatAsLinear());
     ComputeLinearization(&linearization_);
