+++ include/ilqgames/solver/ilq_solver.h
@@ -68,14 +68,54 @@ class ILQSolver : public GameSolver {
   virtual ~ILQSolver() {}
   ILQSolver(const std::shared_ptr<Problem>& problem,
             const SolverParams& params = SolverParams())
-      : GameSolver(problem, params) {}
+      : GameSolver(problem, params) {
+    // Set up LQ solver.
+    if (params_.open_loop)
+      lq_solver_.reset(
+          new LQOpenLoopSolver(problem_->Dynamics(), problem_->NumTimeSteps()));
+    else
+      lq_solver_.reset(
+          new LQFeedbackSolver(problem_->Dynamics(), problem_->NumTimeSteps()));
+  }
+
+  // Solve this game. Returns true if converged.
+  virtual std::shared_ptr<SolverLog> Solve(
+      bool* success = nullptr,
+      Time max_runtime = std::numeric_limits<Time>::infinity());
 
  protected:
-  // Populate the given vector with a linearization of the dynamics about
-  // the given operating point.
-  virtual void ComputeLinearization(
-      const OperatingPoint& op,
-      std::vector<LinearDynamicsApproximation>* linearization);
+  // Modify LQ strategies to improve convergence properties.
+  // This function replaces an Armijo linesearch that would take place in ILQR.
+  // Returns true if successful, and records if we have converged and the total
+  // costs for all players at the new operating point, as well as the times at
+  // which each player achieves an extreme cost.
+  virtual bool ModifyLQStrategies(std::vector<Strategy>* strategies,
+                                  OperatingPoint* current_operating_point,
+                                  bool* is_new_operating_point_feasible,
+                                  bool* has_converged,
+                                  std::vector<float>* total_costs) const;
+
+  // Compute distance (infinity norm) between states in the given dimensions.
+  // If dimensions empty, checks all dimensions.
+  virtual float StateDistance(const VectorXf& x1, const VectorXf& x2,
+                              const std::vector<Dimension>& dims) const;
+
+  // Compute the current operating point based on the current set of strategies
+  // and the last operating point. Checks whether the solver has converged and
+  // populates the total costs for all players of the new operating point.
+  // Returns true if the new operating point satisfies the trust region
+  // (including all explicit inequality constraints), or if the
+  // `check_trust_region` flag is false. Optionally also returns the times of
+  // extreme costs.
+  virtual bool CurrentOperatingPoint(
+      const OperatingPoint& last_operating_point,
+      const std::vector<Strategy>& current_strategies,
+      OperatingPoint* current_operating_point, bool* has_converged,
+      std::vector<float>* total_costs, bool check_trust_region = true,
+      bool* satisfies_constraints = nullptr) const;
+
+  // Core LQ Solver.
+  std::unique_ptr<LQSolver> lq_solver_;
 };  // class ILQSolver
 
 }  // namespace ilqgames
